seperated the pwm lights stuff from ModalLights
